requirements.txt
Django==1.8.3
djangorestframework==3.2.4
ipython==3.2.1
psycopg2==2.6.1
